requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
Django==3.1.7
django-ckeditor==6.0.0
django-form-utils==1.0.3
django-js-asset==1.2.2
django-quill==0.1.0
django-quill-editor==0.1.22
django-tinymce==3.3.0
djangorestframework==3.12.2
et-xmlfile==1.0.1
idna==2.10
openpyxl==3.0.7
PaycomUz==3.0
Pillow==8.2.0
pytz==2021.1
requests==2.25.1
sqlparse==0.4.1
urllib3==1.26.4
xlwt==1.3.0
